Updated README.md
The update explains the large changes that were made. Also the updated read me gives credit where it's due to other team members contributions to the project. Team members are referenced within the read me. Explanation of the project as well as directions on how to setup and run it are publicly available now on the repo. Enjoy!